% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palette-theme.R
\name{theme_volcano}
\alias{theme_volcano}
\title{A refined volcano plot theme}
\usage{
theme_volcano(base_size = 12, base_family = "", legend_position = "top")
}
\arguments{
\item{base_size}{Base font size.}

\item{base_family}{Base font family.}

\item{legend_position}{Legend position passed to \code{\link[ggplot2:theme]{ggplot2::theme()}}.}
}
\value{
A ggplot2 theme.
}
\description{
A refined volcano plot theme
}
